+++ packages/lib-react-components/src/SubjectCard/components/InteractiveMedia/components/Image/Image.jsx
