+++ packages/lib-react-components/src/SubjectCard/components/InteractiveMedia/components/Image/Image.spec.jsx
@@ -1,11 +1,11 @@
 import { render, screen } from '@testing-library/react'
 import sinon from 'sinon'
 
-import SubjectThumbnail from './SubjectThumbnail'
+import Image from './Image'
 
 const src = 'https://panoptes-uploads.zooniverse.org/production/subject_location/66094a64-8823-4314-8ef4-1ee228e49470.jpeg'
 
-describe('SubjectThumbnail', function () {
+describe('Image', function () {
   class ValidImage {
     constructor() {
       this.naturalHeight = 200
@@ -27,27 +27,27 @@ describe('SubjectThumbnail', function () {
   })
 
   it('should render without crashing', async function () {
-    render(<SubjectThumbnail alt='a test image' src={src} />)
+    render(<Image alt='a test image' src={src} />)
     const image = await screen.findByRole('img', { name: 'a test image' })
     expect(image).toBeTruthy()
   })
 
   it('should use alt text to describe the image', async function () {
     const alt = 'A galaxy'
-    render(<SubjectThumbnail alt={alt} src={src} />)
+    render(<Image alt={alt} src={src} />)
     const image = await screen.findByRole('img', { name: alt })
     expect(image).toBeTruthy()
   })
 
   it('should have a `<noscript />` image for SSR', function () {
-    render(<SubjectThumbnail src={src} />)
+    render(<Image src={src} />)
     const noscriptWrapper = document.querySelector('noscript')
     expect(noscriptWrapper).toBeDefined()
   })
 
   describe('height and width', function () {
     it('should be set if specified', async function () {
-      render(<SubjectThumbnail alt='a test image' height={200} width={270} src={src} />)
+      render(<Image alt='a test image' height={200} width={270} src={src} />)
       const image = await screen.findByRole('img', { name: 'a test image' })
       const imageWrapper = screen.getByTestId('subject-thumbnail')
       const { maxHeight, maxWidth } = window.getComputedStyle(imageWrapper)
@@ -56,7 +56,7 @@ describe('SubjectThumbnail', function () {
     })
 
     it('should fill the image container by default', async function () {
-      render(<SubjectThumbnail alt='a test image' src={src} />)
+      render(<Image alt='a test image' src={src} />)
       const image = await screen.findByRole('img', { name: 'a test image' })
       const imageWrapper = screen.getByTestId('subject-thumbnail')
       const { maxHeight, maxWidth } = window.getComputedStyle(imageWrapper)
