+++ packages/lib-react-components/src/SubjectCard/components/InteractiveMedia/components/Image/index.js
@@ -1,0 +1,1 @@
+export { default } from './Image'
