--- packages/lib-react-components/src/SubjectCard/components/InteractiveMedia/components/SubjectThumbnail/index.js
@@ -1,1 +1,0 @@
-export { default } from './SubjectThumbnail'
